aus.fertility {addb}	R Documentation

Australian fertility data

Description

Age-specific fertility rates and female child-bearing population for Australia.

Format

Object of class demogdata containing the following components:

year

Vector of years

age

Vector of ages

rate

List containing one matrix with one age group per row and one column per year.

pop

Population data in same form as rate.

type

Type of object. In this case, “fertility”.

label

Character string giving area from which data are taken. In this case, “Australia”.

Details

Australian fertility rates and populations (1921-2007) for single year age groups (15-49) as provided by the Australian Bureau of Statistics. The population numbers for 2002 onwards differ slighly from those in the australia (mortality) object. The population numbers in aus.fertility have been updated to the latest ABS estimates.

Author(s)

Rob J Hyndman

See Also

aus.fert

Examples

plot(aus.fertility)

[Package addb version 3.222 Index]